Added the I to ICP
est_tform now iteratively estimates the transform and then assigns
nearest detection points. Also returns the points "assigned" to
the model in the last iteration, so that the outer loop can remove
those points from the detections list for the next thing to be
tracked.

NOTE: this is SLOW (several seconds for one model). May have to
drop into C to do this for long videos

Next up: Putting the Multi- in Multi-ICP.
 - loop through the frame calling est_tform and removing the
   detections "under" the aligned frame
 - Note: We'll need a way to ignore 1-2 pixel noise detections