Added the M to Multi-ICP!
Now have the full multi-ICP pipeline! Also, skimage has a transform
estimator which is way faster than the scipy.optimize.min approach,
but which appears to be less stable frame to frame. Left the option
in to use the slower code, but it is disabled by default.

Also added in the boilerplate to run this as a script and make bbox
video. This API will change when btf output becomes the standard
use-case.